🐛 Made publish script flexible

+++ publish.sh
@@ -7,23 +7,32 @@ echo '
             
 For geeks who wanted to write their blog site in Vue               V0.0.1 Alpha
 ================================================================================'
-echo "Removing temporary build directory..."
-rm -rf ./dist
-# git add .
-# git commit -m "Update"
-echo "Building website using `yarn build`..."
-yarn build
-echo "Build completed."
-cd dist
-echo "Initializing Git repository under ./dist..."
-git init
-git add .
-git commit -m "Update [TOOL]"
-echo "Pushing to GitHub..."
-git remote add origin git@github.com:samzhangjy/samzhangjy.github.io.git
-git push origin master --force
-echo "Repository pushed to GitHub."
-cd ..
-echo "Removing temporary build directory..."
-rm -rf ./dist
-echo "Build finished! Your website will be live in a few minutes. Enjoy!"
+echo Enter Github repo link
+read repo_link
+
+if [[ -n $repo_link ]]
+then
+
+  echo "Removing temporary build directory..."
+  rm -rf ./dist
+
+  echo "Building website using `yarn build`..."
+  yarn build
+  echo "Build completed."
+  cd dist
+  echo "Initializing Git repository under ./dist..."
+  git init
+  git add .
+  git commit -m "Update [TOOL]"
+  echo "Pushing to GitHub..."
+  git remote add origin git@github.com:samzhangjy/samzhangjy.github.io.git
+  git push origin master --force
+  echo "Repository pushed to GitHub."
+  cd ..
+  echo "Removing temporary build directory..."
+  rm -rf ./dist
+  echo "Build finished! Your website will be live in a few minutes. Enjoy!"
+
+else 
+  echo "! Repo link is required !"
+fi
